refactor: Cookieのバックアップと復元機能をブラウザモジュールに移動
